+++ tests/test_formatting.py
@@ -4,6 +4,7 @@ import re
 import pytest
 
 from loguru import logger
+from loguru._logger import _Interpolation, _Template
 
 
 @pytest.mark.parametrize(
@@ -266,3 +267,35 @@ def test_invalid_format_key_raises_enhanced_error_without_catch(format_, coloriz
     logger.add(lambda msg: None, format=format_, catch=False, colorize=colorize)
     with pytest.raises(ValueError, match=r"Failed to format log record: key 'missing' not found."):
         logger.opt(colors=colors).info("Hello")
+
+
+@pytest.mark.skipif(_Template is None, reason="Template Strings not supported")
+def test_template_string(writer):
+    # We can't just use t"2**8 = {2**8}", because its a syntax error before python-3.14
+    logger.add(writer)
+    logger.info(_Template("2**8 = ", _Interpolation(2**8)))
+    result = writer.read()
+    assert result.endswith("2**8 = 256\n")
+
+
+@pytest.mark.skipif(_Template is None, reason="Template Strings not supported")
+def test_template_string_is_lazy(writer):
+    # We can't just use t"debug = {debug_tracker}", because its a syntax error before python-3.14
+    class StrCalledTracker:
+        def __init__(self):
+            self.str_called = False
+
+        def __str__(self):
+            self.str_called = True
+            return "xxx"
+
+    logger.add(writer, level="INFO")
+    debug_tracker = StrCalledTracker()
+    info_tracker = StrCalledTracker()
+    logger.debug(_Template("debug = ", _Interpolation(debug_tracker)))  # Should be ignored (debug)
+    logger.info(_Template("info = ", _Interpolation(info_tracker)))  # Should be logged (info)
+    result = writer.read()
+    assert len(result.strip().split("\n")) == 1
+    assert result.endswith("info = xxx\n")
+    assert not debug_tracker.str_called
+    assert info_tracker.str_called